% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_data}
\alias{pg_data}
\title{Download data from Pangaea.}
\usage{
pg_data(doi, path = "~/.pangaea/", overwrite = TRUE, ...)
}
\arguments{
\item{doi}{DOI of Pangaeae single dataset, or of a collection of datasets.}

\item{path}{(character) Path to store files in. Default: \emph{"~/.pangaea/"}}

\item{overwrite}{(logical) Ovewrite a file if one is found with the same name}

\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}
}
\value{
One or more items of class pangaea, each with a citation object, metadata object,
and data object. Each data object is printed as a \code{\link[dplyr]{tbl_df}} object, but the
actual object is simply a \code{data.frame}.
}
\description{
Grabs data as a dataframe or list of dataframes from a Pangaea data repository URI; see:
\url{http://www.pangaea.de/}.
}
\examples{
\donttest{
# a single file
res <- pg_data(doi='10.1594/PANGAEA.807580')
res
res[[1]]$citation
res[[1]]$meta

# another single file
pg_data(doi='10.1594/PANGAEA.807584')

# Many files
res <- pg_data(doi='10.1594/PANGAEA.807587')
res[[3]]

# Another example of many files
res <- pg_data(doi='10.1594/PANGAEA.761032')
res[[1]]
res[[2]]
}
}
\author{
Naupaka Zimmerman
}

