% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_get_data_from_uri}
\alias{pg_get_data_from_uri}
\title{Functions to download data from Pangaea.}
\usage{
pg_get_data_from_uri(uri)
}
\arguments{
\item{uri}{Public internet address of the data}
}
\value{
dataframe with rows of data, list of dataframes with rows of data if there are multiple related data sets found at the URI
}
\description{
Grabs data as a dataframe or list of dataframes from a Pangaea data repository URI; see: \url{http://www.pangaea.de/}. Expects the individual data files to be in tab-delimited text format
}
\examples{
\donttest{
.get_data_from_uri("http://api.example.com/ecologydata")
}
}
\author{
Naupaka Zimmerman
}

