% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_get_record}
\alias{pg_get_record}
\title{Get record from the Pangaea repository}
\usage{
pg_get_record(identifier, prefix = "oai_dc", transform = TRUE)
}
\arguments{
\item{identifier}{Dataset identifier. See Examples.}

\item{transform}{(logical) Transform the output to XML via \code{\link[XML]{xmlParse}}}
}
\description{
Get record from the Pangaea repository
}
\examples{
\donttest{
record <- pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.788382")
record$identifier
record$metadata

record <- pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.269656",
prefix="iso19139")
record$identifier
}
}

