% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_list_sets}
\alias{pg_list_sets}
\title{List the set structure of the Pangaea repository}
\usage{
pg_list_sets(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}
}
\description{
List the set structure of the Pangaea repository
}
\examples{
\donttest{
head( pg_list_sets() )
library('httr')
res <- pg_list_sets(verbose())
}
}

