% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangaear-package.R
\docType{package}
\name{pangaear-package}
\alias{pangaear-package}
\alias{pangaear}
\title{Client for the Pangaea Database}
\description{
\href{https://www.pangaea.de/}{Pangaea database}
}
\details{
Package includes tools to interact with the Pangaea Database,
including functions for searching for data, fetching datasets by
dataset ID, working with the Pangaea OAI-PMH service, and
Elasticsearch service.
}
