% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_get_record.R
\name{pg_get_record}
\alias{pg_get_record}
\title{Get record from the Pangaea repository}
\usage{
pg_get_record(identifier, prefix = "oai_dc", ...)
}
\arguments{
\item{identifier}{Dataset identifier. See Examples.}

\item{prefix}{A character string to specify the metadata format in OAI-PMH requests issued to
the repository. The default (\code{"oai_dc"}) corresponds to the mandatory OAI unqualified
Dublin Core metadata schema.}

\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}
}
\description{
Get record from the Pangaea repository
}
\examples{
\donttest{
record <- pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.788382")
record$header
record$metadata

record <- pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.269656",
prefix="iso19139")
record$header
record$metadata

record <- pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.269656",
prefix="dif")
record$header
record$metadata

# curl options
library('httr')
pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.788382", config=verbose())

# invalid record id
# pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.11111")
# pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.11111", prefix="adfadf")
}
}

