% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{pg_identify}
\alias{pg_identify}
\title{Identify information about the Pangaea repository}
\usage{
pg_identify(transform = TRUE)
}
\arguments{
\item{transform}{A logical indicating whether the OAI-PMH XML results to “useful” R data
structures via oaih_transform. Default: true.}
}
\description{
Identify information about the Pangaea repository
}
\examples{
\dontrun{
pg_identify()
}
}
