% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_identify}
\alias{pg_identify}
\title{Identify information about the Pangaea repository}
\usage{
pg_identify(transform = TRUE)
}
\arguments{
\item{transform}{(logical) Whether the OAI-PMH XML results to \emph{useful} R data
structures via \code{link{oaih_transform}}. Default: \code{TRUE}.}
}
\description{
Identify information about the Pangaea repository
}
\examples{
pg_identify()
}

