% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_identify}
\alias{pg_identify}
\title{Identify information about the Pangaea repository}
\usage{
pg_identify(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}

\item{x}{Input to print.}
}
\description{
Identify information about the Pangaea repository
}
\examples{
\donttest{
pg_identify()
}
}

