% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_identify.R
\name{pg_identify}
\alias{pg_identify}
\title{Identify information about the Pangaea repository}
\usage{
pg_identify(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}
}
\value{
list
}
\description{
Identify information about the Pangaea repository
}
\examples{
\dontrun{
pg_identify()
}
}

