% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_identify.R
\name{pg_identify}
\alias{pg_identify}
\title{Identify information about the Pangaea repository}
\usage{
pg_identify(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{\link[httr:GET]{httr::GET()}}}
}
\value{
list
}
\description{
Identify information about the Pangaea repository
}
\examples{
\dontrun{
pg_identify()
}
}
\references{
\href{https://www.openarchives.org/pmh/}{OAI-PMH documentation}
}
\seealso{
wraps \code{\link[oai:id]{oai::id()}}

Other oai methods: \code{\link{pg_get_record}},
  \code{\link{pg_list_identifiers}},
  \code{\link{pg_list_metadata_formats}},
  \code{\link{pg_list_records}}, \code{\link{pg_list_sets}}
}
