% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_list_metadata_formats}
\alias{pg_list_metadata_formats}
\title{Get metadata formats from the Pangaea repository}
\usage{
pg_list_metadata_formats(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata formats from the Pangaea repository
}
\examples{
\donttest{
pg_list_metadata_formats()
}
}

