% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_list_metadata_formats.R
\name{pg_list_metadata_formats}
\alias{pg_list_metadata_formats}
\title{Get metadata formats from the Pangaea repository}
\usage{
pg_list_metadata_formats(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}
}
\value{
data.frame
}
\description{
Get metadata formats from the Pangaea repository
}
\examples{
\dontrun{
pg_list_metadata_formats()
}
}

