% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_list_metadata_formats.R
\name{pg_list_metadata_formats}
\alias{pg_list_metadata_formats}
\title{Get metadata formats from the Pangaea repository}
\usage{
pg_list_metadata_formats(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{\link[oai:list_metadataformats]{oai::list_metadataformats()}}}
}
\value{
data.frame
}
\description{
Get metadata formats from the Pangaea repository
}
\examples{
\dontrun{
pg_list_metadata_formats()
}
}
\references{
\href{https://www.openarchives.org/pmh/}{OAI-PMH documentation}
}
\seealso{
wraps \code{\link[oai:list_metadataformats]{oai::list_metadataformats()}}

Other oai methods: \code{\link{pg_get_record}},
  \code{\link{pg_identify}},
  \code{\link{pg_list_identifiers}},
  \code{\link{pg_list_records}}, \code{\link{pg_list_sets}}
}
\concept{oai methods}
