% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_list_records}
\alias{pg_list_records}
\title{List records from Pangaea}
\usage{
pg_list_records(prefix = "oai_dc", from = NULL, until = NULL,
  set = NULL, transform = TRUE)
}
\arguments{
\item{prefix}{A character string to specify the metadata format in OAI-PMH requests issued to
the repository. The default (\code{"oai_dc"}) corresponds to the mandatory OAI unqualified
Dublin Core metadata schema.}

\item{from}{Character string giving datestamp to be used as lower bound for datestamp-based
selective harvesting (i.e., only harvest records with datestamps in the given range). Dates
and times must be encoded using ISO 8601. The trailing Z must be used when including time.
OAI-PMH implies UTC for data/time specifications.}

\item{until}{Character string giving a datestamp to be used as an upper bound,
for datestamp-based selective harvesting (i.e., only harvest records with datestamps in
the given range).}

\item{set}{A character string giving a set to be used for selective harvesting (i.e., only
harvest records in the given set).}

\item{transform}{A logical indicating whether the OAI-PMH XML results to \emph{useful} R data
structures via \code{\link[OAIHarvester]{oaih_transform}}. Default: \code{TRUE}.}
}
\description{
List records from Pangaea
}
\examples{
\donttest{
res <- pg_list_records(from='2012-01-01', until='2012-01-15')
head(res); NROW(res)

res <- pg_list_records(set='geomound', from='2012-01-01', until='2012-01-05')
head(res); NROW(res)

# When no results found > "Error: Received condition 'noRecordsMatch'"
pg_list_records(set='geomound', from='2012-01-01', until='2012-01-01')
}
}

