% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pg_list_sets}
\alias{pg_list_sets}
\title{List the set structure of the Pangaea repository}
\usage{
pg_list_sets(transform = TRUE)
}
\arguments{
\item{transform}{logical; whether the OAI-PMH XML results to \emph{useful} R data
structures via \code{link{oaih_transform}}. Default: \code{TRUE}.}
}
\description{
List the set structure of the Pangaea repository
}
\examples{
\donttest{
pg_list_sets()
}
}

