% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_list_sets.R
\name{pg_list_sets}
\alias{pg_list_sets}
\title{List the set structure of the Pangaea repository}
\usage{
pg_list_sets(token = NULL, as = "df", ...)
}
\arguments{
\item{token}{(character) a token previously provided by the server to resume a
request where it last left off. 50 is max number of records returned. We will
loop for you internally to get all the records you asked for.}

\item{as}{(character) What to return. One of "df" (for data.frame; default),
"list", or "raw" (raw text)}

\item{...}{Curl debugging options passed on to \code{\link[httr]{GET}}}
}
\description{
List the set structure of the Pangaea repository
}
\examples{
\dontrun{
pg_list_sets()
library('httr')
res <- pg_list_sets(config = verbose())
}
}

